refactor(shared-testutils-ui): moved check-input-defaults into testutils lib
