+++ libs/shared/testutils/ui/tsconfig.lib.json
@@ -2,9 +2,7 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "../../../../dist/out-tsc",
-    "types": [
-      "node"
-    ]
+    "types": ["node", "@types/testing-library__jest-dom"]
   },
   "files": [
     "../../../../node_modules/@nrwl/react/typings/cssmodule.d.ts",
@@ -21,10 +19,5 @@
     "src/**/*.spec.jsx",
     "src/**/*.test.jsx"
   ],
-  "include": [
-    "src/**/*.js",
-    "src/**/*.jsx",
-    "src/**/*.ts",
-    "src/**/*.tsx"
-  ]
+  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx"]
 }
